\name{MuViCP-package}
\alias{MuViCP-package}
\alias{MuViCP}

\docType{package}

\title{MuViCP}

\description{MultiClass Visualizable Classification using Combination of Projections}

\details{
This is a novel classifier that natively works as an ensemble. It projects data on a large number of matrices, and uses very simple classifiers on each of these projections. The results are then combined, ideally via Dempster-Shafer Calculus.}

\author{Mohit Dayal}
